adapt app to new RouteResponse and Recommendation model, first simple UI pass

+++ priobike-new/lib/models/user_position.dart
@@ -1,7 +1,7 @@
 class UserPosition {
   double? lat;
   double? lon;
-  int? speed;
+  double? speed;
 
   UserPosition({required this.lat, required this.lon, required this.speed});
